refactor: Move task machinery out of server.ts to task_execution.ts
executeToolAndUpdateTask and emitTaskStatusNotification move verbatim
to src/mcp/task_execution.ts, following the tool_dispatch.ts pattern.
server.ts imports emitTaskStatusNotification back for its two
remaining call sites. server.ts: 1863 -> 1371 lines. Closes #1084.
